Back/api/requirements.txt
Flask==2.3.2
Flask-Cors==5.0.0
psycopg2-binary==2.9.5
requests
python-dotenv